refactor(hmpb): extract `all-bubble` ballot rendering files
